+++ src/GafferScene/IECoreScenePreview/CapturingRenderer.cpp
@@ -50,8 +50,23 @@ using namespace IECoreScenePreview;
 IECoreScenePreview::Renderer::TypeDescription<CapturingRenderer> CapturingRenderer::g_typeDescription( "Capturing" );
 
 CapturingRenderer::CapturingRenderer( RenderType type, const std::string &fileName, const IECore::MessageHandlerPtr &messageHandler )
-	:	m_messageHandler( messageHandler ), m_rendering( false )
+	:	m_messageHandler( messageHandler ), m_renderType( type ), m_rendering( false )
 {
+}
+
+CapturingRenderer::~CapturingRenderer()
+{
+	for( auto o : m_capturedObjects )
+	{
+		// Reset soon-to-be-dangling pointer from CapturedObject, in
+		// case clients keep the object alive longer than the renderer.
+		o.second->m_renderer = nullptr;
+		if( m_renderType != Interactive )
+		{
+			// Remove reference added in `object()`.
+			o.second->removeRef();
+		}
+	}
 }
 
 const CapturingRenderer::CapturedObject *CapturingRenderer::capturedObject( const std::string &name ) const
@@ -90,6 +105,11 @@ Renderer::ObjectInterfacePtr CapturingRenderer::camera( const std::string &name,
 	return this->object( name, camera, attributes );
 }
 
+Renderer::ObjectInterfacePtr CapturingRenderer::camera( const std::string &name, const std::vector<const IECoreScene::Camera *> &samples, const std::vector<float> &times, const AttributesInterface *attributes )
+{
+	return this->object( name, vector<const Object *>( samples.begin(), samples.end() ), times, attributes );
+}
+
 Renderer::ObjectInterfacePtr CapturingRenderer::light( const std::string &name, const IECore::Object *object, const AttributesInterface *attributes )
 {
 	return this->object( name, object, attributes );
@@ -102,7 +122,7 @@ Renderer::ObjectInterfacePtr CapturingRenderer::lightFilter( const std::string &
 
 Renderer::ObjectInterfacePtr CapturingRenderer::object( const std::string &name, const IECore::Object *object, const AttributesInterface *attributes )
 {
-	return this->object( name, { object }, { 0.0f }, attributes );
+	return this->object( name, { object }, {}, attributes );
 }
 
 Renderer::ObjectInterfacePtr CapturingRenderer::object( const std::string &name, const std::vector<const IECore::Object *> &samples, const std::vector<float> &times, const AttributesInterface *attributes )
@@ -134,6 +154,14 @@ Renderer::ObjectInterfacePtr CapturingRenderer::object( const std::string &name,
 	CapturedObjectPtr result = new CapturedObject( this, name, samples, times );
 	result->attributes( attributes );
 	a->second = result.get();
+	if( m_renderType != Interactive )
+	{
+		// For non-interactive renders, the client code will typically drop
+		// their reference to the object immediately, but we still want to
+		// capture it for later examination. Add a reference to keep it alive.
+		// See ~CapturingRenderer for the associated `removeRef()`.
+		a->second->addRef();
+	}
 
 	return result;
 }
@@ -193,7 +221,12 @@ CapturingRenderer::CapturedObject::CapturedObject( CapturingRenderer *renderer, 
 
 CapturingRenderer::CapturedObject::~CapturedObject()
 {
-	m_renderer->m_capturedObjects.erase( m_name );
+	if( m_renderer && m_renderer->m_renderType == RenderType::Interactive )
+	{
+		// If the client of an interactive render drops ownership, that means
+		// they want the object to be deleted from the renderer.
+		m_renderer->m_capturedObjects.erase( m_name );
+	}
 }
 
 const std::vector<IECore::ConstObjectPtr> &CapturingRenderer::CapturedObject::capturedSamples() const