+++ src/GafferScene/RendererAlgo.cpp
@@ -155,7 +155,7 @@ void transformSamples( const ScenePlug *scene, size_t segments, const Imath::V2f
 	}
 }
 
-void objectSamples( const ScenePlug *scene, size_t segments, const Imath::V2f &shutter, std::vector<IECoreScene::ConstVisibleRenderablePtr> &samples, std::vector<float> &sampleTimes )
+void objectSamples( const ScenePlug *scene, size_t segments, const Imath::V2f &shutter, std::vector<IECore::ConstObjectPtr> &samples, std::vector<float> &sampleTimes )
 {
 	samples.clear();
 	sampleTimes.clear();
@@ -165,9 +165,12 @@ void objectSamples( const ScenePlug *scene, size_t segments, const Imath::V2f &s
 	if( !segments )
 	{
 		ConstObjectPtr object = scene->objectPlug()->getValue();
-		if( const VisibleRenderable *renderable = runTimeCast<const VisibleRenderable>( object.get() ) )
+		if(
+			runTimeCast<const VisibleRenderable>( object.get() ) ||
+			runTimeCast<const Camera>( object.get() )
+		)
 		{
-			samples.push_back( renderable );
+			samples.push_back( object.get() );
 		}
 		return;
 	}
@@ -189,7 +192,10 @@ void objectSamples( const ScenePlug *scene, size_t segments, const Imath::V2f &s
 		const MurmurHash objectHash = scene->objectPlug()->hash();
 		ConstObjectPtr object = scene->objectPlug()->getValue( &objectHash );
 
-		if( const Primitive *primitive = runTimeCast<const Primitive>( object.get() ) )
+		if(
+			runTimeCast<const Primitive>( object.get() ) ||
+			runTimeCast<const Camera>( object.get() )
+		)
 		{
 			// We can support multiple samples for these, so check to see
 			// if we actually have something moving.
@@ -197,7 +203,7 @@ void objectSamples( const ScenePlug *scene, size_t segments, const Imath::V2f &s
 			{
 				moving = true;
 			}
-			samples.push_back( primitive );
+			samples.push_back( object.get() );
 			lastHash = objectHash;
 		}
 		else if( runTimeCast<const VisibleRenderable>( object.get() ) )
@@ -221,6 +227,31 @@ void objectSamples( const ScenePlug *scene, size_t segments, const Imath::V2f &s
 	if( !moving )
 	{
 		samples.resize( std::min<size_t>( samples.size(), 1 ) );
+		sampleTimes.clear();
+	}
+}
+
+void objectSamples( const ScenePlug *scene, size_t segments, const Imath::V2f &shutter, std::vector<IECoreScene::ConstVisibleRenderablePtr> &samples, std::vector<float> &sampleTimes )
+{
+	vector<ConstObjectPtr> oSamples;
+	vector<float> oSampleTimes;
+	objectSamples( scene, segments, shutter, oSamples, oSampleTimes );
+
+	samples.clear();
+	for( size_t i = 0; i < oSamples.size(); ++i )
+	{
+		if( auto ts = runTimeCast<const VisibleRenderable>( oSamples[i] ) )
+		{
+			samples.push_back( ts );
+			if( i < oSampleTimes.size() )
+			{
+				sampleTimes.push_back( oSampleTimes[i] );
+			}
+		}
+	}
+
+	if( samples.size() < 2 )
+	{
 		sampleTimes.clear();
 	}
 }
@@ -1090,18 +1121,59 @@ struct CameraOutput : public LocationOutput
 		const size_t cameraMatch = m_cameraSet.match( path );
 		if( cameraMatch & IECore::PathMatcher::ExactMatch )
 		{
-			IECore::ConstObjectPtr object = scene->objectPlug()->getValue();
-			if( const Camera *camera = runTimeCast<const Camera>( object.get() ) )
+			// Sample cameras and apply globals
+
+			vector<ConstObjectPtr> samples; vector<float> sampleTimes;
+			RendererAlgo::objectSamples( scene, deformationSegments(), shutter(), samples, sampleTimes );
+
+			vector<ConstCameraPtr> cameraSamples; cameraSamples.reserve( samples.size() );
+			for( const auto &sample : samples )
 			{
-				IECoreScene::CameraPtr cameraCopy = camera->copy();
+				if( auto cameraSample = runTimeCast<const Camera>( sample.get() ) )
+				{
+					IECoreScene::CameraPtr cameraSampleCopy = cameraSample->copy();
+					GafferScene::RendererAlgo::applyCameraGlobals( cameraSampleCopy.get(), m_globals, scene );
+					cameraSamples.push_back( cameraSampleCopy );
+				}
+			}
 
-				GafferScene::RendererAlgo::applyCameraGlobals( cameraCopy.get(), m_globals, scene );
+			// Create ObjectInterface
 
-				IECoreScenePreview::Renderer::ObjectInterfacePtr objectInterface = renderer()->camera(
-					name( path ),
-					cameraCopy.get(),
-					attributesInterface().get()
+			if( !samples.size() || cameraSamples.size() != samples.size() )
+			{
+				IECore::msg(
+					IECore::Msg::Warning,
+					"RendererAlgo::CameraOutput",
+					boost::format( "Camera missing for location \"%1%\" at frame %2%" )
+						% name( path )
+						% Context::current()->getFrame()
 				);
+			}
+			else
+			{
+				IECoreScenePreview::Renderer::ObjectInterfacePtr objectInterface;
+				if( !sampleTimes.size() )
+				{
+					objectInterface = renderer()->camera(
+						name( path ),
+						cameraSamples[0].get(),
+						attributesInterface().get()
+					);
+				}
+				else
+				{
+					vector<const Camera *> rawCameraSamples; rawCameraSamples.reserve( cameraSamples.size() );
+					for( auto &c : cameraSamples )
+					{
+						rawCameraSamples.push_back( c.get() );
+					}
+					objectInterface = renderer()->camera(
+						name( path ),
+						rawCameraSamples,
+						sampleTimes,
+						attributesInterface().get()
+					);
+				}
 
 				if( objectInterface )
 				{
@@ -1232,7 +1304,7 @@ struct ObjectOutput : public LocationOutput
 			return true;
 		}
 
-		vector<ConstVisibleRenderablePtr> samples; vector<float> sampleTimes;
+		vector<ConstObjectPtr> samples; vector<float> sampleTimes;
 		RendererAlgo::objectSamples( scene, deformationSegments(), shutter(), samples, sampleTimes );
 		if( !samples.size() )
 		{